// sample-app/src/swagger/model/student.ts
/**
 * Api Documentation
 * Api Documentation
 *
 * OpenAPI spec version: 1.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Subject } from "./subject";

export interface Student {
  age?: number;
  clazz?: string;
  id?: number;
  name?: string;
  subjects?: Array<Subject>;
}